tms/requirements/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.5.2
    # via django
django==3.2
    # via
    #   -r requirements.in
    #   djangorestframework
    #   djangorestframework-simplejwt
django-js-asset==2.0.0
    # via django-mptt
django-mptt==0.13.4
    # via -r requirements.in
django-simple-history==3.1.1
    # via -r requirements.in
djangorestframework==3.13.1
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.2.0
    # via -r requirements.in
psycopg2==2.9.3
    # via -r requirements.in
pyjwt==2.4.0
    # via djangorestframework-simplejwt
python-dotenv==0.21.0
    # via -r requirements.in
pytz==2022.2.1
    # via
    #   django
    #   djangorestframework
sqlparse==0.4.2
    # via django
